fix : Loading with BuyDash
